+++ tests/PhpWord/Writer/Word2007/Part/SettingsTest.php
@@ -346,4 +346,20 @@ class SettingsTest extends \PHPUnit\Framework\TestCase
         $element = $doc->getElement($path, $file);
         $this->assertNotEquals('false', $element->getAttribute('w:val'));
     }
+
+    public function testConsecutiveHyphenLimit()
+    {
+        $phpWord = new PhpWord();
+        $phpWord->getSettings()->setConsecutiveHyphenLimit(2);
+
+        $doc = TestHelperDOCX::getDocument($phpWord);
+
+        $file = 'word/settings.xml';
+
+        $path = '/w:settings/w:consecutiveHyphenLimit';
+        $this->assertTrue($doc->elementExists($path, $file));
+
+        $element = $doc->getElement($path, $file);
+        $this->assertSame('2', $element->getAttribute('w:val'));
+    }
 }